Added user search API

+++ controller/postControl.js
@@ -29,12 +29,15 @@ exports.addPost = async (req, res) => {
 
 exports.updatePost = async (req, res) => {
     try {
+        const userId = req.id;
         const { newCaption } = req.body;
         const { postId } = req.params;
         const userPost = await Post.findOne({ _id: postId });
-        userPost.caption = newCaption || userPost.caption
-        userPost.save();
-        res.status(201).json({ message: "Post Updated!" });
+        if (userPost.user === userId) {
+            userPost.caption = newCaption || userPost.caption
+            userPost.save();
+            res.status(201).json({ message: "Post Updated!" });
+        }
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
@@ -44,11 +47,14 @@ exports.deletePost = async (req, res) => {
     try {
         const { postId } = req.params;
         const userId = req.id;
-        const user = await User.findOne({ _id: userId });
-        user.posts = user.posts.filter(post => post.toString() !== postId);
-        user.save()
-        await Post.deleteOne({ _id: postId });
-        res.status(201).json({ message: "Post deleted!" });
+        const userPost = await Post.findOne({ _id: postId })
+        if (userPost.user === userId) {
+            user.posts = user.posts.filter(post => post.toString() !== postId);
+            const user = await User.findOne({ _id: userId });
+            user.save()
+            await Post.deleteOne({ _id: postId });
+            res.status(201).json({ message: "Post deleted!" });
+        }
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
